+++ src/components/User/KycRequests/KycRequest.List.vue
@@ -213,6 +213,7 @@ export default {
       try {
         const chunk = await this.list.fetchNext()
         this.list._data = this.list.data.concat(chunk.data)
+        this.list.fetchNext = chunk.fetchNext
         this.isListEnded = oldLength === this.list.data.length
       } catch (error) {
         ErrorHandler.process(error)